refactor: 용어집과 용어 domain 결합
앞으로의 일을 미리 예단하여 나눠놓으면, 쓸대없는 복잡성 증가로 인해 그 부담을 지속적으로 개발자가 짊어지어야 하는 문제가 있음을 인지 함. 미리 설계 및 계획하지 말고, 지금의 스팩에 가장 어울이고, 쉽고 편하게 개발 할 수 있는 형태를 갖도록 하는 것이 좋다고 판단하여 리팩토링 진행
