--- api/src/main/java/vook/server/api/domain/term/package-info.java
@@ -1,9 +1,0 @@
-@ApplicationModule(
-        type = ApplicationModule.Type.OPEN,
-        allowedDependencies = {
-                "common"
-        }
-)
-package vook.server.api.domain.term;
-
-import org.springframework.modulith.ApplicationModule;
