--- api/src/main/java/vook/server/api/domain/vocabulary/model/TermId.java
@@ -1,14 +1,0 @@
-package vook.server.api.domain.vocabulary.model;
-
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@EqualsAndHashCode
-@NoArgsConstructor
-@AllArgsConstructor
-public class TermId {
-    private Long id;
-}
