+++ api/src/main/java/vook/server/api/domain/vocabulary/service/VocabularyService.java
@@ -7,14 +7,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
-import vook.server.api.domain.vocabulary.exception.TermLimitExceededException;
 import vook.server.api.domain.vocabulary.exception.VocabularyLimitExceededException;
 import vook.server.api.domain.vocabulary.exception.VocabularyNotFoundException;
-import vook.server.api.domain.vocabulary.exception.VocabularyTermNotFoundException;
-import vook.server.api.domain.vocabulary.model.*;
+import vook.server.api.domain.vocabulary.model.UserId;
+import vook.server.api.domain.vocabulary.model.Vocabulary;
+import vook.server.api.domain.vocabulary.model.VocabularyRepository;
 import vook.server.api.domain.vocabulary.service.data.VocabularyCreateCommand;
 import vook.server.api.domain.vocabulary.service.data.VocabularyDeleteCommand;
-import vook.server.api.domain.vocabulary.service.data.VocabularyTermAddCommand;
 import vook.server.api.domain.vocabulary.service.data.VocabularyUpdateCommand;
 
 import java.util.List;
@@ -26,7 +25,6 @@ import java.util.List;
 public class VocabularyService {
 
     private final VocabularyRepository repository;
-    private final VocabularyTermRepository termsRepository;
 
     public List<Vocabulary> findAllBy(@NotNull UserId userId) {
         return repository.findAllByUserId(userId);
@@ -53,20 +51,5 @@ public class VocabularyService {
 
     public Vocabulary getByUid(@NotBlank String vocabularyUid) {
         return repository.findByUid(vocabularyUid).orElseThrow(VocabularyNotFoundException::new);
-    }
-
-    public void addTerm(@Valid VocabularyTermAddCommand command) {
-        Vocabulary vocabulary = repository.findByUid(command.getVocabularyUid()).orElseThrow(VocabularyNotFoundException::new);
-        int termCount = vocabulary.termCount();
-        if (termCount >= 100) {
-            throw new TermLimitExceededException();
-        }
-
-        vocabulary.addTerm(command.getTermId());
-    }
-
-    public Vocabulary getByTermId(@NotNull TermId termId) {
-        VocabularyTerm vocabularyTerm = termsRepository.findByTermId(termId).orElseThrow(VocabularyTermNotFoundException::new);
-        return vocabularyTerm.getVocabulary();
     }
 }
