+++ api/src/main/java/vook/server/api/domain/vocabulary/service/data/TermUpdateCommand.java
@@ -1,16 +1,19 @@
-package vook.server.api.domain.term.service.data;
+package vook.server.api.domain.vocabulary.service.data;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import lombok.Builder;
 import lombok.Getter;
-import vook.server.api.domain.term.model.Term;
+import vook.server.api.domain.vocabulary.model.Term;
 
 import java.util.List;
 
 @Getter
-public class TermCreateCommand {
+public class TermUpdateCommand {
+
+    @NotNull
+    private String uid;
 
     @NotBlank
     @Size(min = 1, max = 100)
@@ -24,8 +27,9 @@ public class TermCreateCommand {
     private List<String> synonyms;
 
     @Builder
-    private static TermCreateCommand of(String term, String meaning, List<String> synonyms) {
-        TermCreateCommand command = new TermCreateCommand();
+    private static TermUpdateCommand of(String uid, String term, String meaning, List<String> synonyms) {
+        TermUpdateCommand command = new TermUpdateCommand();
+        command.uid = uid;
         command.term = term;
         command.meaning = meaning;
         command.synonyms = synonyms;
@@ -33,6 +37,6 @@ public class TermCreateCommand {
     }
 
     public Term toEntity() {
-        return Term.forCreateOf(term, meaning);
+        return Term.forUpdateOf(term, meaning, synonyms);
     }
 }
