+++ api/src/main/java/vook/server/api/usecases/package-info.java
@@ -2,8 +2,7 @@
         type = ApplicationModule.Type.OPEN,
         allowedDependencies = {
                 "user",
-                "vocabulary",
-                "term"
+                "vocabulary"
         }
 )
 package vook.server.api.usecases;
