+++ api/src/main/java/vook/server/api/usecases/term/UpdateTermUseCase.java
@@ -3,14 +3,11 @@ package vook.server.api.usecases.term;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.domain.term.model.Term;
-import vook.server.api.domain.term.service.TermService;
-import vook.server.api.domain.term.service.data.TermUpdateCommand;
 import vook.server.api.domain.user.model.User;
 import vook.server.api.domain.user.service.UserService;
-import vook.server.api.domain.vocabulary.model.TermId;
-import vook.server.api.domain.vocabulary.model.Vocabulary;
-import vook.server.api.domain.vocabulary.service.VocabularyService;
+import vook.server.api.domain.vocabulary.model.Term;
+import vook.server.api.domain.vocabulary.service.TermService;
+import vook.server.api.domain.vocabulary.service.data.TermUpdateCommand;
 import vook.server.api.usecases.common.polices.VocabularyPolicy;
 
 import java.util.List;
@@ -21,15 +18,13 @@ import java.util.List;
 public class UpdateTermUseCase {
 
     private final UserService userService;
-    private final VocabularyService vocabularyService;
     private final VocabularyPolicy vocabularyPolicy;
     private final TermService termService;
 
     public void execute(Command command) {
         User user = userService.getByUid(command.userUid());
         Term term = termService.getByUid(command.termUid());
-        Vocabulary vocabulary = vocabularyService.getByTermId(new TermId(term.getId()));
-        vocabularyPolicy.validateOwner(user, vocabulary);
+        vocabularyPolicy.validateOwner(user, term.getVocabulary());
 
         termService.update(command.toServiceCommand());
     }
