+++ api/src/test/java/vook/server/api/domain/term/service/TermServiceTest.java
@@ -6,8 +6,11 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
-import vook.server.api.domain.term.model.Term;
-import vook.server.api.domain.term.service.data.TermCreateCommand;
+import vook.server.api.domain.user.model.User;
+import vook.server.api.domain.vocabulary.model.Term;
+import vook.server.api.domain.vocabulary.model.Vocabulary;
+import vook.server.api.domain.vocabulary.service.TermService;
+import vook.server.api.domain.vocabulary.service.data.TermCreateCommand;
 import vook.server.api.globalcommon.exception.ParameterValidateException;
 import vook.server.api.testhelper.IntegrationTestBase;
 import vook.server.api.testhelper.creator.TestUserCreator;
@@ -35,7 +38,11 @@ class TermServiceTest extends IntegrationTestBase {
     @DisplayName("용어 생성 - 성공")
     void create() {
         // given
+        User user = userCreator.createCompletedOnboardingUser();
+        Vocabulary vocabulary = vocabularyCreator.createVocabulary(user);
+
         TermCreateCommand command = TermCreateCommand.builder()
+                .vocabularyUid(vocabulary.getUid())
                 .term("용어")
                 .meaning("용어 설명")
                 .synonyms(List.of("동의어1", "동의어2"))
